Added Mouse Wheel Listener -Slider Now movees when mouse wheel is rolled -All components are required to have a visibility variable because I need a a way of showing/hiding components
Next commit is going implmenting is_visible, which is not as simple as it seems because given component 'X' you have to make sure that
any component 'Y' that has 'X' as an ancestor is also invisible.

is_invisible is just means it will get "ignored" by the manager for rendering and event listening.