Implemented invisible components - works by looping through all components/nodes and crawling up to the root to check if a node has been disabled - the results are stored in a table (Vec<bool>) and are recomputed every frame - event listerners do not get fired for invisible components,however, I can think of cases where bugs can occur.
Imagine the scenerio where...
1) mouse hoverIn's to a red component(lets call it X) X changes white when hoverIn listener is handled and if
it were to hoverOut we would go back to being red
2) something makes X invisible, mouse moves out of X
3) something makkes X visible, mouse is outside of X
4) X is white when the user intended for it to be red when the mouse is not hovering the component

Before I set something visible I should enqueue a symetric event meaning
I should check if something is on hoverIN and send hoverOut before making it invisible.

The problem is that this causes problems with the interface.

Every component has is_visible:bool somewhere in the state. So it makes sense for a user to modify that,however,
in order to implement the solution above, I'd have to CHECK the current state of what is being hovered/clicked/dragged and that
is stored in the GuiManager, with the thre pointers clicked_component,hover_component,and focused_component.

I'd have to write a GuiManager::set_visibility(&mut self,v:bool)

Which ,again, would be confusing.

Maybe I should rename GuiComponent::visibility(..) to GuiComponent::local_visibility(..)