Added the Glutin backend
Notes:
-I'm going to make glutin the default windowing backend for this library
because it links to native libraries directly. I'm not 100% certain but
I think on linux it talks directly to Wayland and Xorg like how my
lib "Xproto" does,however, it has a glutin has a superior interface and
way more functionality.
-Unlike sdl2 Glutin is strictly a windowing library, so I'll have to
figure out how to push audio to the system on my own.
-I'll probably just use libpulse for linux, and WASAPI for windows
-in order to select SDL2 just say cargo build -feature=sdl2