Updated Readme + Working on ALSA backend
I'm mainly making this commit to update the readme to better explain
the project, what it does, and how to use it. I think ALSA is default
on all/most linux distros so i don't need go any further than that. Linux
has a SHITLOAD of audio subsystems and libraries but I think ALSA is
good enough.