Prepping for WASAPI implementation
In previous commit (regarding ALSA implemtation) I actually fixed the
problem with sleeping. The Problem was that I was taking the lock
of the mutex and then SLEEPING the thread (huge L). Anyway I caught that
and am now moving onto WASAPI.

I'm getting ready to add the WASAPI implementation, I put the ALSA
implementation in its own module,likewise,the WASAPI implementation goes
in its own module.

Note on CrossCompilation:
When building my examples on WINE with the glutin backend, I noticed that
window dimensions requested is not what the final window dimesntions
actually are. In the brick_demo I request an 800x600 window, however,
i get a window something like 1000x700 or something, which fucked up my
demo and I had to mod the demo to deal with this new behavior in WINE.

I guess I have to make clear that a lot of settings that are specified in
the config file are merely suggestions under WINE and it could do
unexpected things...