WASAPI implemented
It just pumps the audio data from the user back to the sound card like
ALSA, however, There was an issue with mono that I temporarily corrected,
where it only plays in one speaker. This doesn't happen when you
select mono for WEBAUDIO,SDL2 or ALSA. The temp fix was to just manually
mix the mono data to stereo, but I'm not sure I should have to do that
explicitly. I suspect there may be a way for me to mix mono=>stereo
in WASAPI.