Added Clippy Linter (part 1)
I found out about clippy and decided to finally use it. It appearently
helps you write more readable rust code so I've hooked it up my
language server. I'm currently refactoring the entire project according
to clippy's suggestions. This is going to be a big refactor so I'm doing
it in parts.