Moved modules
I moved the modules:
-ogl
-text_writer

to the root of the crate, because I don't want the "extras" a default
feature anymore and while those two modules ARE optional they are
relatively small and don't rely on other, larger crates.

Everything else in "extras" can now be completely optional so I can
avoid pulling in and compiling the codec Lewton and PureMP3.