Did resampling on WASAPI
Turns out you can't just request whatever samplerate you want on WASAPI,
rather, the API gives you a sample rate and it is MY job to resample
to the the incoming audio to that frequency. So I've implemented basic
resampling and have bumped up buffering from 100ms to 200ms. 200ms
is interactive enough for my use-case but the latency could be lacking
for some people. Maybe I'll make buffering length a tweakable parameter
in the future.